% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract}
\alias{extract}
\title{Extract values from dataset.}
\usage{
extract(data, pattern, period, scol = "acc", cn = NULL, by = "mb")
}
\arguments{
\item{data}{Data frame with budget data.}

\item{pattern}{Regular expression to match.}

\item{period}{Years to be extracted.}

\item{scol}{Column name where \code{pattern} is searched for: default is \code{acc} for key in the chart of accounts, alternative is \code{desc} for account description.}

\item{cn}{If set, values are aggregated over \code{period}.}

\item{by}{Grouping variable; default is \code{mb}.}
}
\value{
Data frame with grouping variable and extracted values (aggregated if \code{cn} is set).
}
\description{
Extract values from dataset.
}
\note{
Arbitrary values of \code{scol} and \code{by} are ok they exist in \code{data}.
}
