% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.budget.R
\name{import.budget}
\alias{import.budget}
\title{Import budget data from raw Excel file.}
\usage{
import.budget(path, rm.sheets = 1:5, rm.cols = c(5, 7, 9), skip = 9)
}
\arguments{
\item{path}{Path to file.}

\item{rm.sheets}{Sheets to remove; if NULL, all sheets are imported.}

\item{rm.cols}{Columns to remove.}

\item{skip}{Number of rows to skip.}
}
\description{
Import budget data from raw Excel file.
}
\details{
All sheets will be imported except those specified in \code{rm.sheets}. Sheet name is used as value of \code{lau.name} and first two digits of file name are used as value of \code{county.code} in the resulting data frame.
}
