% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jlprsBudget.2010.2014.R
\docType{data}
\name{jlprsBudget.2010.2014}
\alias{jlprsBudget.2010.2014}
\title{jlprsBudget dataset (2010-2014).}
\format{A data frame with 578796 rows and 10 variables:
\describe{
  \item{mb}{jlprs UID.}
  \item{county.code}{county code.}
  \item{lau.name}{lau.name.}
  \item{acc}{accounting scheme ID.}
  \item{desc}{accounting scheme description.}
  \item{y2010}{data for 2010.}
  \item{y2011}{data for 2011.}
  \item{y2012}{data for 2012.}
  \item{y2013}{data for 2013.}
  \item{y2014}{data for 2014.}
}}
\source{
\url{http://www.mfin.hr/hr/ostvarenje-proracuna-jlprs-za-period-2010-2014/}
}
\usage{
jlprsBudget.2010.2014
}
\description{
jlprsBudget dataset (2010-2014).
}
\keyword{datasets}
