% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jlprsBudget.2014.2017.R
\docType{data}
\name{jlprsBudget.2014.2017}
\alias{jlprsBudget.2014.2017}
\title{jlprsBudget dataset (2014-2017).}
\format{A data frame with 619384 rows and 9 variables:
\describe{
  \item{mb}{jlprs UID.}
  \item{county.code}{county code.}
  \item{lau.name}{lau.name.}
  \item{acc}{accounting scheme ID.}
  \item{desc}{accounting scheme description.}
  \item{y2014}{data for 2014.}
  \item{y2015}{data for 2015.}
  \item{y2016}{data for 2016.}
  \item{y2017}{data for 2017.}
}}
\source{
\url{http://www.mfin.hr/hr/ostvarenje-proracuna-jlprs-za-period-2014-2017/}
}
\usage{
jlprsBudget.2014.2017
}
\description{
jlprsBudget dataset (2014-2017).
}
\keyword{datasets}
